# frontend/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DailyTools247 API",
    "description": "Comprehensive set of free utility and AI-powered APIs for developers, including text analysis, security checkers, formatting utilities, social media copy generation, and email marketing analytics.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://www.dailytools247.app/api",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/text/word-counter": {
      "post": {
        "summary": "Count words and analyze text statistics",
        "description": "Calculates word count, character count, sentence count, paragraph count, and estimated reading time.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "The text to analyze."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "words": { "type": "integer" },
                        "characters": { "type": "integer" },
                        "sentences": { "type": "integer" },
                        "paragraphs": { "type": "integer" },
                        "readingTimeMinutes": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/text/case-converter": {
      "post": {
        "summary": "Convert text case",
        "description": "Converts input text to upper, lower, title, sentence, camel, pascal, snake, kebab, alternating, or inverse case.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string" },
                  "case": {
                    "type": "string",
                    "enum": ["upper", "lower", "title", "sentence", "camel", "pascal", "snake", "kebab", "alternating", "inverse"],
                    "default": "upper"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "converted": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/text/markdown-to-html": {
      "post": {
        "summary": "Convert Markdown to HTML",
        "description": "Translates standard Markdown text elements (headings, bold, italic, code blocks, links, lists) into clean HTML nodes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["markdown"],
                "properties": {
                  "markdown": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "html": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/text/ai-text-summarizer": {
      "post": {
        "summary": "Summarize text using AI",
        "description": "Summarizes long paragraphs of text into a concise summary with a specified number of sentences.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string" },
                  "sentences": { "type": "integer", "default": 3 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "summary": { "type": "string" },
                        "originalSentences": { "type": "integer" },
                        "summarySentences": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/social/ai-hashtag-generator": {
      "post": {
        "summary": "Generate hashtags using AI",
        "description": "Creates highly relevant, high-reach hashtags based on keywords and selected platform.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["keywords"],
                "properties": {
                  "keywords": { "type": "string" },
                  "platform": { "type": "string", "enum": ["instagram", "tiktok", "twitter", "linkedin"], "default": "instagram" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "hashtags": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/social/ai-bio-generator": {
      "post": {
        "summary": "Generate social media profile bios using AI",
        "description": "Creates customized biographies and descriptions for social accounts based on occupation and tone.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "role", "keywords"],
                "properties": {
                  "name": { "type": "string" },
                  "role": { "type": "string" },
                  "keywords": { "type": "string" },
                  "tone": { "type": "string", "default": "professional" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "bios": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}